# docs/swagger.yaml
openapi: 3.0.3
info:
  title: ZoikoShield - Enterprise Cyber Security Platform API
  version: 1.0.0
  description: >
    Complete OpenAPI specification for ZoikoShield Platform Services: Shield Core, Ingest, Anchor, Action, and AI microservices.
    Auth is primary cookie-based (httpOnly access_token / refresh_token, Secure, SameSite=Lax) or Bearer token for API clients / OAuth2.
servers:
  - url: http://localhost:3001
    description: Local development (Shield Core Gateway)
  - url: http://localhost:3002
    description: Ingest Microservice
  - url: http://localhost:3003
    description: Anchor Microservice
  - url: http://localhost:3004
    description: Action Microservice
  - url: http://localhost:3005
    description: AI Gateway Microservice
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: Primary auth mechanism — httpOnly, Secure, SameSite=Lax. Set automatically by login/verify-email/refresh.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer extractor for API clients, OAuth2 tokens, and inter-service calls.
    InternalServiceAuth:
      type: apiKey
      in: header
      name: x-internal-service-token
      description: >
        Shared-secret header (INTERNAL_SERVICE_TOKEN) required by InternalAuthGuard.
        Restricted to service-to-service calls between shield-core/shield-ai/shield-action/shield-anchor —
        never issued to customers or developer-API clients.
...
  schemas:
...
    CreateServiceObligationRequest:
      type: object
      required: [contractId, description]
      properties:
        contractId: { type: string, format: uuid }
        description: { type: string }
        dueAt: { type: string, format: date-time }
...
    CreateDraftInvoiceRequest:
      type: object
      required: [commercialAccountId, amount]
      properties:
        commercialAccountId: { type: string, format: uuid }
        amount: { type: number }
        currency: { type: string, example: "USD" }
    CreateCatalogVersionRequest:
      type: object
      required: [versionName]
      properties:
        versionName: { type: string, example: "2026.1" }
    CreateProductRequest:
      type: object
      required: [sku, name]
      properties:
        sku: { type: string, example: "SHIELD-ENT-001" }
        name: { type: string }
    CreatePriceBookRequest:
      type: object
      required: [sku, unitPrice]
      properties:
        sku: { type: string }
        unitPrice: { type: number }
        currency: { type: string }
...
paths:
...
  /api/v1/obligations:
    post:
      tags: [ServiceObligations]
      summary: Create service obligation
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CreateServiceObligationRequest" } } }
      responses:
        "201": { description: Obligation created }
    get:
      tags: [ServiceObligations]
      summary: Get obligations by contract ID
      parameters: [{ name: contractId, in: query, required: true, schema: { type: string } }]
      responses:
        "200": { description: Obligations list }
  /api/v1/obligations/{id}/status:
    patch:
      tags: [ServiceObligations]
      summary: Update obligation status & link evidence
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: Obligation status updated }
...
  /api/v1/billing/invoices:
    post:
      tags: [Billing]
      summary: Create draft invoice
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CreateDraftInvoiceRequest" } } }
      responses:
        "201": { description: Invoice created }
    get:
      tags: [Billing]
      summary: List account invoices
      parameters: [{ name: accountId, in: query, required: true, schema: { type: string } }]
      responses:
        "200": { description: Invoices list }
  /api/v1/billing/invoices/{id}/issue:
    patch:
      tags: [Billing]
      summary: Issue invoice (FIN-02 lock)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: Invoice issued and locked }
  /api/v1/catalog/versions:
    post:
      tags: [Catalog]
      summary: Create catalog version
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CreateCatalogVersionRequest" } } }
      responses:
        "201": { description: Catalog version created }
  /api/v1/catalog/products:
    post:
      tags: [Catalog]
      summary: Register product SKU
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CreateProductRequest" } } }
      responses:
        "201": { description: Product registered }
    get:
      tags: [Catalog]
      summary: List approved catalog products
      responses:
        "200": { description: Product list }

  # ── Developer API & OAuth ──
...
  /api/v1/catalog/versions/{id}/approve:
    patch:
      tags: [Catalog]
      summary: Approve a catalog version
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: Version approved }
  /api/v1/catalog/price-books:
    post:
      tags: [Catalog]
      summary: Create a price book
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CreatePriceBookRequest" } } }
      responses:
        "201": { description: Price book created }
  /api/v1/catalog/price-books/{id}/approve:
    patch:
      tags: [Catalog]
      summary: Approve a price book
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: Price book approved }
...